requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
charset-normalizer==2.0.12
colorama==0.4.4
discord-py-interactions==4.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.28
orjson==3.6.7
protobuf==3.19.4
pycodestyle==2.8.0
python-decouple==3.6
toml==0.10.2
yarl==1.7.2